Temporary workaround to grant Storage permission.
We have logic that automatically grants the legacy "Storage"
permission when the user grants either of the new "Music" or
"Photos & Videos" permissions, but that logic only lives in
PermissionController, and does not apply to pre-grants.

As a temporary workaround, explicitly grant the "Storage"
permission in pre-grants until we can refactor the
PermissionController logic.

Bug: 124466734
Test: manual
Change-Id: Ic3b07e82092a9100aea89263a1175d9fe2284382
(cherry picked from commit 3805e94129eacfb04c776bff4de308238d5d28fc)