Set with local device info when TV sends out <System Audio Mode Request> with the current device's physical address
When we receive <System Audio Mode Request> from TV with a new Active
Path, we check if this path matches any device info in the local device
list. But the local device itself is not in the list. So it's possible
that we set Active Source as an unregistered device with the same
physical address as ourselves. This patch fixes the issue.

Test: manual
Bug: 131629012
Change-Id: I74180f86c0e905507d6262e13c783a43f7e25adf