Greylist 3P packages in the bootclasspath.
These packages contain code that was whitelisted in P (accidentally),
and was blacklisted in Q (until now). They also contain open source code
that apps could reasonably include in their APK. Blacklisting the code
is problematic, as this results in apps being able to resolve the class
itself, but without any visibility of any members within it.

This is a measure to reduce the risk of Q, pending a proper solution for
these classes in future.

Remove the equivalent methods from the greylist as they are now
redundant. This was done via:
$ cat hiddenapi-greylist.txt \
    | grep -v '^Lgov/nist/' \
    | grep -v '^Lorg/apache/harmony/' \
    | grep -v '^Lorg/apache/xalan/' \
    | grep -v '^Lorg/apache/xml/' \
    | grep -v '^Lorg/apache/xpath/'

Bug: 129387816
Test: m
Change-Id: I48d4ca3b8069894e91785f1dd69d81878078208c