Pass FalsingManager down through more classes.
This is the first step in getting the code to inject FalsingManager
into more places via Dagger.

This removes FalsingManagerFactory, replacing it with
calls to Depedency.get() and pushes those calls further
up the stack in hopes that the FalsingManager can be directly
injected where it's needed in the future.

Contains a few random lint cleanups.

Bug: 136279712
Test: atest SystemUITests
Change-Id: Ie24acd33b0d81ae2112aead0c0d5a8658e04994e