Fix back-swipe gesture on bubble expanded state
Tapping around the edges of the expanded state collapses the bubble, it
also means back-swipe doesn't work properly / show the animation (even
though the bubble is technically collapsing).

This CL includes a bit of a slop around ActivityView when checking if
the touch is intersecting it. We might need to adjust the value but
in my testing so far it seems much better.

Also removes onInterceptTouchEvent code -- that was totally unneeded!!

Fixes: 131267438
Test: manual - expand a bubble and tap near the top / left / right / bottom
               edges => ensure it doesn't collapse when you wouldn't
               expect it to
             - perform backswipe gesture => ensure you see arrow animation
               and bubble collapses

Change-Id: Id42e6d5fbaff832b9d8f0a415d8f91a5358af5de