Stop tracking the secure setting because we're in master
For the QPR we needed to move to global setting, to not break CTS we had
some weird code to keep both global & secure setting in sync (ag/8708041).

This is no longer needed in R/master so this CL removes those code paths
and updates the tests to rely on the global setting.

Test: atest NotificationManagerTest NotificationManagerServiceTest
Fixes: 138398470
Change-Id: Iebb6d59baf2ec3c90816c3f9ba2d1b2e1aa8675a