Make MemoryTile share after Keyguard
If the device is locked clicking on the MemoryTile will take a heapdump
but not prompt for sharing. If the device is locked after a Reboot,
clicking on the MemoryTile will crash the device as the resolver cannot
find an activity.

This makes sure that the Keyguard is dismissed before trying to share.
Using the systemui ActivityStarter also catches any errors.

Fixes: 141701312
Test: manual with device locked and after reboot
Change-Id: I5f0e83f9c95c9b3b33318303d853299422233261